feat: add mask
