+++ packages/core/src/2d/sprite/index.ts
@@ -1,3 +1,5 @@
+export type { IMaskRenderable } from "./MaskRenderable";
+export { MaskDirtyFlags, MaskRenderable } from "./MaskRenderable";
 export { Sprite } from "./Sprite";
 export { SpriteMask } from "./SpriteMask";
 export { SpriteMaskUtils } from "./SpriteMaskUtils";
